% Auto-generated: do not edit by hand
\name{twash}

\alias{twash}

\title{Twash component}

\description{
Twash, A Twitter Profile Timeline component for Dash from musicfox.io.  This component uses react-twitter-widgets under the hood.
}

\usage{
twash(id=NULL, sourceType=NULL, handle=NULL, height=NULL, width=NULL, theme=NULL,
linkColor=NULL, borderColor=NULL, noHeader=NULL, noFooter=NULL, noBorders=NULL,
noScrollbar=NULL, transparent=NULL, lang=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks}

\item{sourceType}{Character. String 'profile' is all for now}

\item{handle}{Character. String user name or public handle, e.g. ArianaGrande}

\item{height}{Numeric. Integer height (in pixels)}

\item{width}{Numeric. Integer width (in pixels)}

\item{theme}{A value equal to: 'dark', 'light'. * With dark or light theme}

\item{linkColor}{Character. * With custom link colors. Note: Only Hex colors are supported.}

\item{borderColor}{Character. * With custom border colors. Note: Only Hex colors are supported.}

\item{noHeader}{Logical. * Hide the header from timeline}

\item{noFooter}{Logical. * Hide the footer from timeline}

\item{noBorders}{Logical. * Hide the border from timeline}

\item{noScrollbar}{Logical. * Hide the scrollbars}

\item{transparent}{Logical. * Enable Transparancy}

\item{lang}{Character. * Custom language code. Supported codes here: https://developer.twitter.com/en/docs/twitter-for-websites/twitter-for-websites-supported-languages/overview.html}
}
